Update addAppCommand to create a new Appointment
A new appointment needs to be made to be added.

Creating the appointment in addAppCommand allows us to use the list
in the model to obtain the client.

Let's use the index given to obtain the Person object that represents
the client and along with the other fields, create an Appointment.

What needs to be done:
* Update the model to use the Schedule class
* Modify the execute method to properly add the appointment to the
Schedule in model.